suport in neovim 0.9+
due to the cscope has been out from neovim 0.9. therefore
it has to leverage dhananjaylatkar/cscope_maps.nvim.

adjust the cscope db add/rm corespoding command to dhananjaylatkar/cscope_maps.nvim
if the editor is nvim and over 0.9

+++ plugin/cs-mgmt.vim
@@ -273,11 +273,19 @@ endif
 " Cscoope's Function"{{{
 if g:CsMgmtCscopeEnable == 1
     func! s:cm_cscope_attach(db) dict
-        exec printf("cs add %s.out", a:db)
+        if has('nvim-9')
+            exec printf("Cs db add %s.out", a:db)
+        else
+            exec printf("cs add %s.out", a:db)
+        endif
     endf
 
     func! s:cm_cscope_detach(db) dict
-        exec printf("cs kill %s.out", a:db)
+        if has('nvim-9')
+            exec printf("Cs db rm %s.out", a:db)
+        else
+            exec printf("cs kill %s.out", a:db)
+        endif
     endf
 
     func! s:cm_cscope_db_build(ref_name) dict
